requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
av==14.0.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
executing==2.1.0
Flask==3.1.0
flatbuffers==24.3.25
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
livekit==0.18.3
livekit-agents==0.12.2
livekit-api==0.8.0
livekit-plugins-openai==0.10.10
livekit-plugins-silero==0.7.4
livekit-protocol==0.7.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.0
onnxruntime==1.20.1
openai==1.57.4
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==4.5.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
watchfiles==0.24.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
